refactored file structure to MVC architecture and decoupled some event logic
