+++ src/js/app.js
@@ -1,28 +1,30 @@
-import { state } from "./state.js";
-import { pubsub } from "./pubsub.js";
-import { renderListTemplate } from "./renderlist.js";
-import { toggleDarkTheme } from "./eventhandler.js";
-import { note } from "./factories.js";
-import { drawCanvas } from "./canvas.js";
-import { initPiano } from "./piano.js";
+import { state } from "./models/state.js";
+import { pubsub } from "./models/pubsub.js";
+import { renderListTemplate } from "./views/renderlist.js";
+import { toggleDarkTheme } from "./controllers/themehandler.js";
+import { note } from "./models/factories.js";
+import { drawCanvas } from "./views/canvas.js";
+import { initPiano } from "./views/piano.js";
+import { useColor } from './controllers/colorhandler';
 
 state.notes.push(
-  new note({ interval: 1, octave: 0, velocity: 1, color: [255, 0, 0] })
+  new note({ interval: 1, octave: 0, velocity: 1, color: [] })
 );
 state.notes.push(
   new note({
     interval: 1.067,
     octave: 0,
     velocity: 0.8,
-    color: [137, 215, 220],
+    color: [],
   })
 );
 
 renderListTemplate(state);
 window.onload = (() => {
   drawCanvas();
+  useColor(state);
+  initPiano();
 })
-initPiano();
 
 
 pubsub.subscribe("note added", renderListTemplate);
@@ -31,3 +33,4 @@ pubsub.subscribe("note added", drawCanvas);
 pubsub.subscribe("note deleted", drawCanvas);
 pubsub.subscribe("controls changed", drawCanvas)
 pubsub.subscribe("theme changed", toggleDarkTheme);
+pubsub.subscribe("color changed", useColor);
