+++ src/js/controllers/controlhandler.js
@@ -1,10 +1,8 @@
-import { state } from "./state.js";
-import { pubsub } from "./pubsub.js";
-import { note } from "./factories.js";
-// import { plane } from "./curtains.mjs";
+import { pubsub } from '../models/pubsub'
+import { state } from '../models/state'
+import { note } from '../models/factories'
 
-  const themeToggle = document.querySelector(".toggle");
-  themeToggle.addEventListener("click", (e) => useTheme(e));
+
 function attachListeners() {
   const params = document.querySelectorAll(".param");
 
@@ -39,20 +37,6 @@ function useParam(e) {
   }
 }
 
-let darkMode = false;
-function useTheme(e) {
-  const parameter = e.currentTarget.getAttribute("data-parameter");
-  if (parameter == "theme") {
-    darkMode = !darkMode;
-    if (darkMode == true) {
-      state.theme = "dark";
-    } else {
-      state.theme = "light";
-    }
-    pubsub.publish("theme changed", state);
-  }
-}
-
 function updateControlsUI(e) {
   const index = e.target.getAttribute("data-index");
   const parameter = e.target.getAttribute("data-parameter");
@@ -78,27 +62,4 @@ function addNote(e) {
   pubsub.publish("note added", state);
 }
 
-function toggleDarkTheme() {
-  const root = document.querySelector(".root");
-  const switcher = document.querySelector(".switch");
-  const icon = document.querySelector(".icon");
-  const mask = document.querySelector(".mask");
-
-  if (state.theme == "dark") {
-    root.style.setProperty("--bg-main", "var(--gray-900)");
-    root.style.setProperty("--bg-focus", "var(--gray-800)");
-    root.style.setProperty("--mid", "var(--gray-500)");
-    root.style.setProperty("--high-contrast", "var(--gray-100)");
-  } else {
-    root.style.setProperty("--bg-main", "var(--gray-100)");
-    root.style.setProperty("--bg-focus", "var(--gray-500)");
-    root.style.setProperty("--mid", "var(--gray-900)");
-    root.style.setProperty("--high-contrast", "var(--gray-800)");
-  }
-  switcher.classList.toggle("switch-left");
-  switcher.classList.toggle("switch-right");
-  mask.classList.toggle("nomask");
-  icon.classList.toggle("sun");
-}
-
-export { attachListeners, toggleDarkTheme };
+export {attachListeners}
