+++ src/js/models/factories.js
@@ -1,8 +1,6 @@
-const note = function ({ interval, octave, velocity, color }) {
+export const note = function ({ interval, octave, velocity, color }) {
   this.interval = interval;
   this.octave = octave;
   this.velocity = velocity;
   this.color = color;
 };
-
-export { note };
