+++ src/js/models/state.js
@@ -1,6 +1,5 @@
 export let state = {
   transpose: 0,
   notes: [],
-  shaderUniforms: {},
   theme: "light",
 };
